// MyMusic.Client/src/model/previewDeviceNamingTemplateSong.ts
// @ts-nocheck
/**
Generated by orval 🍺
Do not edit manually.
MyMusic.Server | v1
OpenAPI spec version: 1.0.0
*/
export interface PreviewDeviceNamingTemplateSong {
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  songDeviceId: number;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  songId: number;
  title: string;
  currentPath: string;
  newPath: string;
  changed: boolean;
  /** @nullable */
  error?: string | null;
}